chore(mobile): accounts related logs
