+++ suite-common/logger/src/logsMiddleware.ts
@@ -1,44 +1,10 @@
 import { isAnyOf } from '@reduxjs/toolkit';
 
 import { createMiddleware } from '@suite-common/redux-utils';
-import {
-    accountsActions,
-    blockchainActions,
-    fiatRatesActions,
-    getFiatStaleTickersThunk,
-    initFiatRatesThunk,
-    onUpdateFiatRateThunk,
-    updateCurrentFiatRatesThunk,
-    updateLastWeekFiatRatesThunk,
-    updateStaleFiatRatesThunk,
-    updateTxsFiatRatesThunk,
-} from '@suite-common/wallet-core';
+import { accountsActions, blockchainActions } from '@suite-common/wallet-core';
 import { getAccountIdentifier } from '@suite-common/wallet-utils';
 
 import { addLog } from './logsSlice';
-
-const isAnyOfFiatRatesActions = isAnyOf(
-    fiatRatesActions.removeFiatRate,
-    fiatRatesActions.updateFiatRate,
-    fiatRatesActions.updateLastWeekFiatRates,
-    fiatRatesActions.updateTransactionFiatRate,
-    getFiatStaleTickersThunk.pending,
-    getFiatStaleTickersThunk.fulfilled,
-    updateCurrentFiatRatesThunk.pending,
-    updateCurrentFiatRatesThunk.fulfilled,
-    updateStaleFiatRatesThunk.pending,
-    updateStaleFiatRatesThunk.fulfilled,
-    onUpdateFiatRateThunk.pending,
-    onUpdateFiatRateThunk.fulfilled,
-    updateLastWeekFiatRatesThunk.pending,
-    updateLastWeekFiatRatesThunk.fulfilled,
-    updateTxsFiatRatesThunk.pending,
-    updateTxsFiatRatesThunk.fulfilled,
-    updateTxsFiatRatesThunk.pending,
-    updateTxsFiatRatesThunk.fulfilled,
-    initFiatRatesThunk.pending,
-    initFiatRatesThunk.fulfilled,
-);
 
 export const logsMiddleware = createMiddleware((action, { next, dispatch }) => {
     if (isAnyOf(accountsActions.createAccount, accountsActions.updateAccount)(action)) {
@@ -71,10 +37,6 @@ export const logsMiddleware = createMiddleware((action, { next, dispatch }) => {
 
     if (blockchainActions.setBackend.match(action)) {
         dispatch(addLog({ type: action.type, payload: { ...action.payload, urls: undefined } }));
-    }
-
-    if (isAnyOfFiatRatesActions(action)) {
-        dispatch(addLog({ type: action.type, payload: { ...action.payload } }));
     }
 
     return next(action);
