+++ suite-native/logger/tsconfig.json
@@ -6,6 +6,9 @@
         {
             "path": "../../suite-common/redux-utils"
         },
+        {
+            "path": "../../suite-common/wallet-core"
+        },
         { "path": "../module-settings" }
     ]
 }
